requirements.txt
matplotlib==3.3.4
numpy==1.19.3
Pillow==8.4.0
torch==1.9.0
torchvision==0.10.0
